Allow instance to launch w/subnet default ipv6 setting
Right now, setting this to 0 launches the instance -- and then causes
the instance to be rebuilt on the next run if the subnet defaults to
assigning ipv6 addresses.

Conditionalizing this on -1 w/null allows the current behaviour to be
preserved, while not clobbering instances on subnets w/ipv6 by default.