Allow instance profiles to be provided
Sometimes we want to use our own instance profiles.  Why?  Well...  I'm
sure you just thought of three or four while reading this paragraph.

Here we allow one to be specified, and make instance profile and role
creation conditional.  We also update our outputs to add an instance
profile output (no sense forcing future-$self to "guess") and ensure
that the role name returned is either the one we created or the one
attached to the provided profile.