Add lifecycle block with ignore_changes
In this case ignoring user_data or user_data_base64. If you share a generic user data between several instances and then make any change to that, it will require recreating the ec2 instance.

Ideally this would be passed in via a variable and configurable, but I do not believe Terraform is capable of doing this today.

I think this is an ok compromise for the current situation.